simplify publish - hide Publication from library users
